Added some code for personalised Spotify API based on kim_diep user
Added some code for personalised Spotify API based on kim_diep user

rspotify wrapper: as a backup plan for demo purposes

spotify-client wrapper: as a backup plan for demo purposes

I added some code in of my own which goes beyond the rspotify and spotify-client wrappers. This is because it was too restricted, so using access-token and oauth, I was able to access my personalised list of top artists and tracks based on HTTP request utilising net/http , uri and json.